raiz-Backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==1.10.24
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
typing_extensions==4.15.0
Werkzeug==3.1.3